PP-10041 Refactor InviteRouter#routeOtpDispatch
Refactor the InviteRouter#routeOtpDispatch to accept an InviteEntity
rather than a code. Instead, look up the InviteEntity in the resource.
This simplifies the code as InviteRouter#routeOtpDispatch no longer
needs to return a Pair with a boolean indicating whether or not the
telephone number and password should be expected in the HTTP request
body.