PP-10048 Refactor exceptions from NotificationService
When the Notify client returns an exception, wrap the exception in a
UserNotificationException and explicitly handle this in calling methods.
Previously, we wrapped the exception in a WebApplicationException which
extends RuntimeException and in the calling method were catching all
Exceptions.

This change makes it easier to understand when exceptions might be
thrown.