PP-10041: Get rid of new InviteType enum constants
We added new InviteType enum constants to give better separation for the three journeys, but overlooked the fact that when a user is created, they might still have outstanding invites (for other services) that were created beforehand, and the types for those invites will then be incorrect.  So, it's not viable to segregate the journeys into three different types at the outset, and we need to get rid of the new enum values and go back to using the two original ones.