PP-10255 Check if there are uncommitted changes to OpenApi file
Add a check in the GitHub actions workflow that runs on PR builds to
check whether there are uncommitted changes after running
`mvn compile`. If there are, this suggests that there have been
changes to the API specification and the OpenApi file has not been
regenerated. Fail the build when uncommitted changes are detected.