PP-10260 Return POJO from resource methods
Instead of constructing and returning a `Response` in resource methods, it is clearer to return
a POJO which will be magically serialised by Dropwizard using Jackson. For error responses, throw
a WebApplicationException which is serialised to an error response.

Change some response codes from 200 -> 204 when no response body is returned. Selfservice handles
all success codes in the same way so no updates to the consumer are required.

Only continue returning a `Response` when we want to return a 201 response rather than 200 when a
resource is created.