PP-10260 Remove pattern that returns exception
Remove the pattern in InviteService where some methods return Optional<WebApplicationException> and
later throws the exception in the resource. It looks like this pattern was used as in
`InviteService#validateOtp` we want to throw an exception but also commit a database update to the
invite if it needs to be disabled without rolling back the exception.

Update the `@Transactional` annotation to be
`@Transactional(ignore = {WebApplicationException.class})` so that we can throw
WebApplicationExceptions without rolling back the transaction.

Change the resource method to return `void` so that Jackson will automatically construct the
response. This changes the status code for a success response from `200` to `204`. `204` is the
correct status code when the response has no body. This change will have no effect on the consumer,
as selfservice treats all success status codes the same when calling this endpoint.