PP-10260 Refactor complete invite resource
Accept a POJO in the method signature for deserialising the request body.

Refactor InviteCompleter to remove the `withData` method and instead pass in the
CompleteInviteRequest as an argument of `InviteCompleter#complete` to make it easier to understand.
The `CompleteInviteRequest` will be null if the request was made with no body.