PP-10260 Remove request body for complete endpoint
The /complete endpoint accepts an optional request body containing a
list of gateway account ids to attach to the service created for a
self-signup.

However, the selfservice code has been modified so a request body is
never provided. Instead, we now link the gateway account to the service
in a separate PATCH request after the service has been created. So this
request no longer needs to accept a request body.