PP-10260 Refactor resend OTP and validate OTP resources
Refactor the resources for validating and resending OTP codes for
invites to use recommended Dropwizard standards for deserialising and
validating the request.

Deserialise the request using a POJO with the appropriate annotations.
Use the @Valid annotation with appropriate annotations to validate
mandatory fields and the validity of provided values.

Make some returned status codes more correct, such as return 204 - No
content when no body is returned for the response.