PP-10244 Do not create a service for a self-registration invite
Previously, as part of the self-registration journey we created a test
service which we asked the user to name. We no longer want to do this.
This was done when the `/v1/api/invites/{code}/complete` endpoint was
called.

Stop creating a service when this endpoint is called. In the response
body, don't return a `service_external_id` for this type of invite.
pay-selfservice already handles this field being absent, and will not
ask the user to enter a service name if it is.

Combine the code for completing invites into one method on
`InviteService`. This previously lived in implementations of the
`InviteCompleter` interface.