PP-11265 add unique index to lower email on the users table
The current sql unique constraint on the email column in the users table is not case insensitive. It should be.

Unique constraint lower(email) is not valid PSQL script. It was either unique index or installing citext extension and altering the column type. citext can only be installed by superuser and cannot be done through our usual liquibase script.

- Add unique index to lower(email).