BAU Return success response when agreement exists
When a request was made to create either a GovUkPayAgreement or
StripeAgreement was made, we were returning a 409 error if one already
existed.

Instead return a 200. For creating a GovUkPayAgreement, also return the
agreement in the response.

This is so that we can recover from the scenario where the end user
submits the admin tool page to agree to the agreement, but there is
another error, which we have seen happen when trying to create the
Zendesk ticket.