BAU fix dropwizard versioning
dependabot will not update dropwizard dependecies because it uses ${dropwizard.version}.
dropwizard-testing-junit4 is also using ${drowizard.version} but it's a different module and its version numbering is out of sync with the main dropwizard dependencies.
We need the updates for dropwizard dependencies not to rely on io.dropwizard.modules versions for security updates.

- separate the 2 dependencies that each of them have their own versiong.