PP-13030: Use CreateServiceRequest pojo instead of JsonNode
Following the general pattern in most of our Dropwizard codebases, we want to avoid
JsonNode as the POST request body where possible so we can advantage of using
javax Validation constraints.

It is also generally agreed that using hard coded URL strings makes for
easier-to-read code.

Added a few more test cases in ServiceResourceCreateIT to reflect the pact
tests between selfservice and adminusers:
* [a valid create service request with empty object](https://github.com/alphagov/pay-selfservice/blob/b2053a1049bc28813e8e7363245ad546ef19b675/test/pact/adminusers-client/service/create-service.pact.test.js#L50)
* [a valid create service request with gateway account ids](https://github.com/alphagov/pay-selfservice/blob/b2053a1049bc28813e8e7363245ad546ef19b675/test/pact/adminusers-client/service/create-service.pact.test.js#L85C31-L85C86)
* [a valid create service request with service name](https://github.com/alphagov/pay-selfservice/blob/b2053a1049bc28813e8e7363245ad546ef19b675/test/pact/adminusers-client/service/create-service.pact.test.js#L124)