PP-13030: Use enums for the role name
When creating a new target service which will only have a Worldpay test account
and is branched off from an original source service, we want to ensure the
admin users on the source service are also admins on the new target service. To
this end we want to get admin users for the existing service. We can get all
users for a service using `GET /v1/api/services/{serviceExternalId}/users`, but
we'd like to introduce a query parameter of role=admin for example. Doing this
has the advantage of transferring less data across the wire, as the User object
is quite large. To this end this commit makes the role an enumerated type which
will provide validation on the role query parameter.

This commit:
* Introduces a RoleName enum
* Deletes the RoleDbFixture. Roles are [initialised into the
  database](https://github.com/alphagov/pay-adminusers/blob/master/src/main/resources/config/initdata/roles.csv),
  so where there are integration tests, there is never a need to create custom
  roles. We should use the predefined roles.
* Deletes the RoleEntityFixture for the same reason as above.

This PR does not:
* Add a role query parameter to GET /v1/api/services/{serviceExternalId}/users
* Enum-ify the role description.