PP-13030: Add is_worldpay_test_service column to services table
We want to set up a separate service specifically for worldpay testing if the
service requires one. This introduces the following functionality:
* There is a "WORLDPAY TEST SERVICE" label shown against the service name in
  the My Services view in the selfservice UI
* This new service is solely for testing, and will not be able to go live

To this end, introducing a is_worldpay_test_service column is the most
straightforward and easy-to-understand way to enable the above functionalities
easily.

This commit also sneakily changes SerivceResource.findService to getService as
this method doesn't find a service by parameters, but gets a service by its
unique; and it's better to follow REST naming conventions, ie. "GET" in this
case.