PP-13647 update AWS SDK to version 2
The AWS SDK for Java 1.x is in maintenance mode, effective July 31, 2024.
The AWS SDK for Java 1.x will no longer receive updates for new or existing services.
We need to upgrade `pay-adminuers` to use the AWS SDK version 2 for SQS operations. This will involve updating dependencies, refactoring code to accommodate changes in the SDK, and ensuring the new SDK's methods and patterns are followed.